fix delete_file/1 in Upload.File
This just fixes a simple edge case of attempting to delete a firmware file stored locally
that doesn't exist anymore. Really, this only affects running NervesHub locally (or anyone
else who might use this file store instead of S3). In the case of local development, fw
files are stored in tmp locations so a restart could blow it away while the record exists
in the app. Then attempting to delete the record would fail. This just checks if it is there
first before attempting to hard delete

+++ apps/nerves_hub_web_core/lib/nerves_hub_web_core/firmwares/upload/file.ex
@@ -24,14 +24,14 @@ defmodule NervesHubWebCore.Firmwares.Upload.File do
 
   @spec delete_file(Firmware.t()) :: :ok
   def delete_file(%{upload_metadata: %{local_path: path}}) do
-    File.rm!(path)
+    # Sometimes fw files may be stored in temporary places that
+    # get cleared on reboots, especially when using this locally.
+    # So if the file doesn't exist, don't attempt to remove
+    if File.exists?(path), do: File.rm!(path), else: :ok
   end
 
-  @spec delete_file(Firmware.t()) ::
-          {:ok, any()}
-          | {:error, any()}
   def delete_file(%{upload_metadata: %{"local_path" => path}}) do
-    File.rm!(path)
+    delete_file(%{upload_metadata: %{local_path: path}})
   end
 
   @spec metadata(Org.id(), String.t()) :: upload_metadata()
